feat(cayenne): add memory accounting for off-pool resident state
